refactor: view 역할 분리
